feat(exception): improve error handling and response messages for exceptions
feat(jwt): enhance JWT error responses with JSON format and status codes

feat(product): update product update method parameter order and change stockQuantity type to Integer

+++ src/main/java/com/adham/store_management_system/dto/ProductResponseDto.java
@@ -16,5 +16,5 @@ public class ProductResponseDto {
     private String productName;
     private BigDecimal productPrice;
     private String categoryName;
-    private double stockQuantity;
+    private Integer stockQuantity;
 }
